Basic Upgrades
This reverts commit 750cdc8179d8ac81e8fb5c5aa7867a722728e168.

+++ Assets/_Scripts/PlayerInfo.cs
@@ -37,7 +37,7 @@ public class PlayerInfo : MonoBehaviour {
 		gameTime = 0.0f;	
 	}
 
-	public void AssignValues(int _playerNo,string _playerName, int _lives, int _maxhp, int _baseSpeed, int _meshtype, int _tiletype,int _colorType, Snake _player, float _starttime){
+	public void AssignValues(int _playerNo,string _playerName, int _lives, int _maxhp, float _baseSpeed, int _meshtype, int _tiletype,int _colorType, Snake _player, float _starttime){
 		PlayerNo = _playerNo;
 		PlayerName = _playerName;
 		Lives = _lives;
